Making tags page consistent with Repositories and Organizations list page
