+++ src/routes/RepositoryDetails/Tags/Tags.tsx
@@ -1,11 +1,8 @@
-import {Toolbar} from './Toolbar';
+import {TagsToolbar} from './TagsToolbar';
 import Table from './Table';
 import {useState, useEffect} from 'react';
-import {
-  filterState,
-  paginationState,
-  selectedTagsState,
-} from 'src/atoms/TagListState';
+import {filterState, selectedTagsState} from 'src/atoms/TagListState';
+import {Page, PageSection, PageSectionVariants} from '@patternfly/react-core';
 import {useRecoilValue, useResetRecoilState} from 'recoil';
 import {
   Tag,
@@ -23,7 +20,6 @@ import {CubesIcon} from '@patternfly/react-icons';
 export default function Tags(props: TagsProps) {
   const [tags, setTags] = useState<Tag[]>([]);
   const filter = useRecoilValue(filterState);
-  const pagination = useRecoilValue(paginationState);
   const [loading, setLoading] = useState<boolean>(true);
   const [err, setErr] = useState<string>();
   const resetSelectedTags = useResetRecoilState(selectedTagsState);
@@ -31,9 +27,12 @@ export default function Tags(props: TagsProps) {
   const filteredTags: Tag[] =
     filter !== '' ? tags.filter((tag) => tag.name.includes(filter)) : tags;
 
+  // Pagination related states
+  const [perPage, setPerPage] = useState<number>(10);
+  const [page, setPage] = useState<number>(1);
   const paginatedTags: Tag[] = filteredTags.slice(
-    (pagination.page - 1) * pagination.perPage,
-    pagination.page * pagination.perPage,
+    (page - 1) * perPage,
+    page * perPage,
   );
 
   const loadTags = async () => {
@@ -91,25 +90,32 @@ export default function Tags(props: TagsProps) {
   }
 
   return (
-    <>
-      <Toolbar
-        organization={props.organization}
-        repository={props.repository}
-        tagCount={filteredTags.length}
-        loadTags={loadTags}
-      ></Toolbar>
-      <ErrorBoundary
-        hasError={isErrorString(err)}
-        fallback={<RequestError message={err} />}
-      >
-        <Table
-          org={props.organization}
-          repo={props.repository}
-          tags={paginatedTags}
-          loading={loading}
-        />
-      </ErrorBoundary>
-    </>
+    <Page>
+      <PageSection variant={PageSectionVariants.light}>
+        <TagsToolbar
+          organization={props.organization}
+          repository={props.repository}
+          tagCount={filteredTags.length}
+          loadTags={loadTags}
+          TagList={paginatedTags}
+          perPage={perPage}
+          page={page}
+          setPage={setPage}
+          setPerPage={setPerPage}
+        ></TagsToolbar>
+        <ErrorBoundary
+          hasError={isErrorString(err)}
+          fallback={<RequestError message={err} />}
+        >
+          <Table
+            org={props.organization}
+            repo={props.repository}
+            tags={paginatedTags}
+            loading={loading}
+          />
+        </ErrorBoundary>
+      </PageSection>
+    </Page>
   );
 }
